% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_utils.R
\name{card_rank}
\alias{card_rank}
\title{Extract Rank from a Card Vector}
\usage{
card_rank(x)
}
\arguments{
\item{x}{A \code{card} vector created using \code{card()}.}
}
\value{
A character vector of ranks ("2"–"10", "J", "Q", "K", "A").
}
\description{
Returns the rank (face value) of each card in a \code{card} vector.
}
\details{
This function returns character values from the \code{rank} field of each card.
Input must be of class \code{card}; otherwise, an error is raised.
}
\examples{
cards <- card(c("A", "10", "Q"), c("♠", "♥", "♦"))
card_rank(cards)  # "A" "10" "Q"

}
