% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_hand.R
\name{deal_hand}
\alias{deal_hand}
\title{Deal a Hand of Blackjack Cards}
\usage{
deal_hand(deck, n = 2)
}
\arguments{
\item{deck}{A character vector of cards (default: deck_cards())}

\item{n}{Number of cards to deal (default is 2)}
}
\value{
A character vector containing the dealt cards
}
\description{
Randomly selects a specified number of cards from a given deck.
This function is typically used to deal an initial hand to a player or dealer
in a round of Blackjack.
}
\examples{
deal_hand(deck_cards()$cards, 2)
}
