% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_class.R
\name{vec_proxy.card}
\alias{vec_proxy.card}
\title{Expose the record data frame underlying a card}
\usage{
vec_proxy.card(x)
}
\arguments{
\item{x}{A 'card' object}
}
\value{
A tibble/data.frame with columns rank and suit
}
\description{
Expose the record data frame underlying a card
}
