% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_class.R
\name{vec_restore.card}
\alias{vec_restore.card}
\title{Repair a card vector after a vctrs operation (subsetting, etc.)}
\usage{
vec_restore.card(x, to, ...)
}
\arguments{
\item{x}{A proxied data frame or record of fields rank + suit}

\item{to}{A prototype 'card' vector (ignored here)}
}
\value{
A new 'card' object with the same rank + suit fields as x
}
\description{
Repair a card vector after a vctrs operation (subsetting, etc.)
}
