% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_player_dealer.R
\name{blackjack_score}
\alias{blackjack_score}
\title{Score a Blackjack Hand}
\usage{
blackjack_score(hand)
}
\arguments{
\item{hand}{A character vector of cards (e.g., c("A♠", "10♦")).}
}
\value{
A numeric value representing the score of the hand.
}
\description{
This function calculates the total score of a given Blackjack hand.
Aces are treated as 11 unless it causes the hand to bust, in which case
they are treated as 1.
}
\examples{
blackjack_score(c("A♠", "10♦"))  # returns 21
blackjack_score(c("A♠", "9♠", "2♦"))  # returns 12
}
