% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_class.R
\name{c.card}
\alias{c.card}
\title{Combine card vectors with c()}
\usage{
\method{c}{card}(..., recursive = FALSE)
}
\arguments{
\item{...}{One or more objects of class 'card'}

\item{recursive}{(ignored)}
}
\value{
A single 'card' vector containing all inputs
}
\description{
Combine card vectors with c()
}
