% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_class.R
\name{c.card}
\alias{c.card}
\title{Combine Multiple Card Vectors}
\usage{
\method{c}{card}(...)
}
\arguments{
\item{...}{One or more objects of class \code{"card"}.}
}
\value{
A combined \code{card} vector.
}
\description{
Method for \code{c()} that combines one or more \code{card} vectors into a single vector.
}
\examples{
hand1 <- card(c("A", "10"), c("♠", "♦"))
hand2 <- card(c("Q"), c("♥"))
c(hand1, hand2)

}
