% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_utils.R
\name{card_rank}
\alias{card_rank}
\title{Extract the rank (value) of a card}
\usage{
card_rank(x)
}
\arguments{
\item{x}{A \code{card} vector.}
}
\value{
A character vector of ranks ("2", "3", …, "10", "J", "Q", "K", "A").
}
\description{
Extract the rank (value) of a card
}
\examples{
cards <- card(c("A", "10", "Q"), c("♠", "♥", "♦"))
card_rank(cards)
}
