% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_utils.R
\name{is_face_card}
\alias{is_face_card}
\title{Is this card a face card (J, Q, or K)?}
\usage{
is_face_card(x)
}
\arguments{
\item{x}{A \code{card} vector.}
}
\value{
A logical vector (TRUE if rank is "J", "Q", or "K").
}
\description{
Is this card a face card (J, Q, or K)?
}
\examples{
cards <- card(c("J", "5", "Q"), c("♣", "♦", "♥"))
is_face_card(cards)  # TRUE, FALSE, TRUE
}
