% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_blackjack_game.R
\name{print.blackjack_game}
\alias{print.blackjack_game}
\title{Print Method for blackjack_game Objects}
\usage{
\method{print}{blackjack_game}(x, ...)
}
\arguments{
\item{x}{An object of class 'blackjack_game'.}

\item{...}{Additional arguments (ignored).}
}
\description{
Nicely formats the output of a simulated Blackjack game.
}
